Modularize autocommand implementation
